--- a/fn/go.mod
+++ b/fn/go.mod
@@ -1,6 +1,6 @@
 module github.com/lightningnetwork/lnd/fn/v2
 
-go 1.19
+go 1.23
 
 require (
 	github.com/stretchr/testify v1.8.1