Haven't had an excuse to make an [iterator](https://pkg.go.dev/iter) yet, so I nerd-sniped myself into the creation of this PR. 

This PR does a few things:
  * Updates nodes+chan updates in horizon to return an iterator of the contents.
     * Even though most of the actual values we read out will already be in the channel graph cache, as is we can have a potentially very long running database transactions. 
     * The new iterator versions accept a batch size and read in chunks, yielding out the responses. Callers are mostly unchanged. 
  *  Optimizes public node filtering for nodes in horizon:
     * Before in a new loop after we'd read all the nodes, we'd then check again to see which ones are public. This previously meant a new DB transaction for each node. Now we fold that into the underlying query. 
  * Update the cache incrementally during reading:
     * Before we'd hold the cache mutex the entire time while reading out the entire response. 
     * Now we'll only hold the cache to check if we can serve from it, then we update the cache items at the very end of serving a batch. 

The cache changes now mean that an invocation doesn't have a consistent view of the cache, but for cases like this (serving gossip data to peers, can be lossy), we don't really need a consistent snapshot. This change should reduce over all mutex contention as well.