@morehouse 2025-08-06
Cool idea.  Will we be able to thread this through to `ApplyGossipFilter` to reduce the memory utilization?

There would still be an outgoing bandwidth issue, but this would help with excessive memory use.