@Roasbeef 2025-08-14
> Cool idea. Will we be able to thread this through to ApplyGossipFilter to reduce the memory utilization?

Yep, I realize now that I stopped just short of threading it up to that level. 

The missing change here would be updating `UpdatesInHorizon` to return either a multi-iterator, or a single one that unifies the two iterator streams. So rather than return `[]lnwire.Message`, it would be `iter.Seq[lnwire.Message]`. Then that main goroutine launched just continues to iterate over the response as normal (need to change that length check w/ something like a one time call to `next()` to see if anything is returned). 

> There would still be an outgoing bandwidth issue, but this would help with excessive memory use.

Isn't this effectively addressed via the `rate.Limiter` usage based on outbound bandwidth sent? It _is_ the case that we don't apply that to broadcasts like we do for the gossip filter application, but gossip filter application is where most of the bandwidth usage comes from, as some implementations effectively request the entire graph on connection.