@morehouse 2025-09-03
> The missing change here would be updating UpdatesInHorizon to return either a multi-iterator, or a single one that unifies the two iterator streams. So rather than return []lnwire.Message, it would be iter.Seq[lnwire.Message]. Then that main goroutine launched just continues to iterate over the response as normal (need to change that length check w/ something like a one time call to next() to see if anything is returned). The callers stay the same for the most part, but now everything is lazy loaded in the background, as needed.

This all sounds great to me.

> Isn't this effectively addressed via the `rate.Limiter` usage based on outbound bandwidth sent? It _is_ the case that we don't apply that to broadcasts like we do for the gossip filter application, but gossip filter application is where most of the bandwidth usage comes from, as some implementations effectively request the entire graph on connection.

My main concern was that one peer could easily monopolize the outbound bandwidth since we would still send the requested gossip as fast as possible.  But  https://github.com/lightningnetwork/lnd/pull/10103 was merged yesterday, which I think addresses this concern.