@Roasbeef 2025-09-16
Pushed a new rebased version. 

Working on some performance tests to show the impact of this PR.