@Roasbeef 2025-09-24
Alright, I've finished my experiments!

### Serial Processing in Master

This was one of those classic cases where I created the experiment assuming one outcome, but gained more insight and arrived at anther outcome in the process. 

The first relevant tidbit is that today in master, we have a mutex that guards access to `ChanUpdatesInHorizon`: https://github.com/lightningnetwork/lnd/blob/82f77e542f569b2eec9e63653de6d10f4e0a3315/graph/db/kv_store.go#L2042-L2055. 

What this means is that only a _single_ goroutine/peer can actually read the backlog from disk at a time. So even though the default filter sema value is 5, only one goroutine can actually read at a time. 

As an aside, I first tried to just remove that mutex to see how it faried against my testing program, but it wasn't pretty: master would OOM in around 5 seconds flat. If it didn't OOM, it would just lock up aggressively trying to GC as the main had no swap configured. 

### Lazy Concurrent Processing w/ this PR

Before I arrived at the above insight, my assumption was that this PR would lead to both lower CPU and memory utilization. That was _nearly_ the case, but not as much as I envisioned. The hidden truth here was that: this PR is able to serve concurrent client (up to the filter-sema) with lower CPU _and_ memory than master, which can only serve a single client a a time. This branch was able to serve the same amount of clients, but with lower mutex contention, less goroutines, and less total allocations. I can provide these profiles if y'all are curious, but the next section paints a better picture w.r.t the gain of this PR. 

### Time-To-First-Message Experiment 

In my experiment, I increased the filter sema value to 200, and ran with 50 concurrent clients. Each client would send a gossip filter, then wait until the first gossip message was received, record the latency, then repeat. I ran this for 5 minutes total. Master was able to process 437 iterations, while this PR could process 4768 iterations (10x improvement). For reference the box I was using has vCPUs, and my internet was a bit slow/wonky at times, so improvign both would likely see a greater improvement. 

I generated a scatterplot, histogram, and CDF with the raw data. I used `gonum` to generate these graphs. 

### `master` 


First master: 

<img width="768" height="384" alt="plain-master_histogram" src="https://github.com/user-attachments/assets/0a915f78-55d2-4592-a642-f4a3f6d57c61" />

<img width="960" height="384" alt="plain-master_scatterplot" src="https://github.com/user-attachments/assets/1d21a0b5-82b5-4fe4-ba4c-d2c56617d8f0" />


<img width="768" height="384" alt="plain-master_cdf" src="https://github.com/user-attachments/assets/6d93d35e-0e80-4801-8a15-976e777e8d30" />

----

The time here is in ms. Unsurprisingly, it can take nearly 70 seconds (!) to respond to a given client, as we process everything serially due to the mutex. 

Raw stats from my testing program: 
```
2025-09-23 17:07:03.492 [INF] GSSP: Total samples: 437
2025-09-23 17:07:03.492 [INF] GSSP: Min response time: 1041ms
2025-09-23 17:07:03.492 [INF] GSSP: Max response time: 72655ms
2025-09-23 17:07:03.492 [INF] GSSP: Average response time: 32485.2ms
2025-09-23 17:07:03.492 [INF] GSSP: Median response time: 29636.0ms
2025-09-23 17:07:03.492 [INF] GSSP: 95th percentile: 71140.0ms
2025-09-23 17:07:03.492 [INF] GSSP: 99th percentile: 72132.0ms
```

### `iter-chan-updates`

Now this branch: 
<img width="960" height="384" alt="iterator_scatterplot" src="https://github.com/user-attachments/assets/36186e1a-ec9f-41f9-83f0-7f1054230a65" />


<img width="768" height="384" alt="iterator_cdf" src="https://github.com/user-attachments/assets/eb7a1afd-5ba1-4c69-b12b-cd9a287ed5c4" />

<img width="768" height="384" alt="iterator_histogram" src="https://github.com/user-attachments/assets/24cdbe4c-c431-4622-8290-2bd5fda61ad1" />

---- 

Compared to master, we can see that we have a much nicer CDF, and also histogram, pretty much everything. The 50th percentile is 500ms, compared to nearly 3 seconds for master: 
```
2025-09-23 17:17:46.044 [INF] GSSP: === Response Time Statistics ===
2025-09-23 17:17:46.044 [INF] GSSP: Total samples: 4768
2025-09-23 17:17:46.044 [INF] GSSP: Min response time: 192ms
2025-09-23 17:17:46.044 [INF] GSSP: Max response time: 3252ms
2025-09-23 17:17:46.044 [INF] GSSP: Average response time: 566.9ms
2025-09-23 17:17:46.044 [INF] GSSP: Median response time: 506.0ms
2025-09-23 17:17:46.044 [INF] GSSP: 95th percentile: 1103.0ms
2025-09-23 17:17:46.044 [INF] GSSP: 99th percentile: 1739.0ms
```


# Conclusion 

This branch is able to: reduce client waiting latency (thereby reducing total CPU allocated), and process concurrent clients with the same or less resources. This iterator pattern should be used in other RPC and p2p instances, where we can lazily stream/send a respond to a given client/peer/user.