@Roasbeef 2025-09-24
You can also see some interesting interaction in the scatter plot for master: each client needs to wait on the next to release the mutex, so the waiting time steadily increases.